--- ghost/adapter-cache-memory-ttl/index.js
@@ -1,1 +1,0 @@
-module.exports = require('./lib/AdapterCacheMemoryTTL');
