--- ghost/adapter-cache-memory-ttl/package.json
@@ -1,28 +1,0 @@
-{
-    "name": "@tryghost/adapter-cache-memory-ttl",
-    "version": "0.0.0",
-    "repository": "https://github.com/TryGhost/Ghost/tree/main/packages/adapter-cache-memory-ttl",
-    "author": "Ghost Foundation",
-    "private": true,
-    "main": "index.js",
-    "scripts": {
-        "dev": "echo \"Implement me!\"",
-        "test:unit": "NODE_ENV=testing c8 --all --check-coverage --reporter text --reporter cobertura -- mocha --reporter dot './test/**/*.test.js'",
-        "test": "yarn test:unit",
-        "lint:code": "eslint *.js lib/ --ext .js --cache",
-        "lint": "yarn lint:code && yarn lint:test",
-        "lint:test": "eslint -c test/.eslintrc.js test/ --ext .js --cache"
-    },
-    "files": [
-        "index.js",
-        "lib"
-    ],
-    "devDependencies": {
-        "c8": "8.0.1",
-        "mocha": "10.2.0",
-        "sinon": "15.2.0"
-    },
-    "dependencies": {
-        "@isaacs/ttlcache": "1.4.1"
-    }
-}
